ajustado metodo remover pessoa, criado metodo remover lancamento

+++ src/main/java/com/gsdev/gsmoney/api/resource/PessoaResource.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -58,6 +59,10 @@ public class PessoaResource {
     @DeleteMapping("/{codigo}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void remover(@PathVariable Long codigo) {
+        if (!pessoaRepository.existsById(codigo)) {
+            throw new EmptyResultDataAccessException(1);
+        }
+
         pessoaRepository.deleteById(codigo);
     }
